--- a/cmd/clickhouse-backup/main.go
+++ b/cmd/clickhouse-backup/main.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	stdlog "log"
 	"os"
+	"runtime"
 	"strconv"
 	"strings"
 
@@ -437,6 +438,10 @@ func main() {
 			UsageText: "clickhouse-backup restore  [-t, --tables=<db>.<table>] [-m, --restore-database-mapping=<originDB>:<targetDB>[,<...>]] [--tm, --restore-table-mapping=<originTable>:<targetTable>[,<...>]] [--partitions=<partitions_names>] [-s, --schema] [-d, --data] [--rm, --drop] [-i, --ignore-dependencies] [--rbac] [--configs] [--named-collections] [--resume] <backup_name>",
 			Action: func(c *cli.Context) error {
 				b := backup.NewBackuper(config.GetConfigFromCli(c))
+				// Override config with CLI flag if provided
+				if c.Bool("restore-in-place") {
+					b.SetRestoreInPlace(true)
+				}
 				return b.Restore(c.Args().First(), c.String("tables"), c.StringSlice("restore-database-mapping"), c.StringSlice("restore-table-mapping"), c.StringSlice("partitions"), c.StringSlice("skip-projections"), c.Bool("schema"), c.Bool("data"), c.Bool("drop"), c.Bool("ignore-dependencies"), c.Bool("rbac"), c.Bool("rbac-only"), c.Bool("configs"), c.Bool("configs-only"), c.Bool("named-collections"), c.Bool("named-collections-only"), c.Bool("resume"), c.Bool("restore-schema-as-attach"), c.Bool("replicated-copy-to-detached"), version, c.Int("command-id"))
 			},
 			Flags: append(cliapp.Flags,
@@ -536,6 +541,11 @@ func main() {
 					Hidden: false,
 					Usage:  "Copy data to detached folder for Replicated*MergeTree tables but skip ATTACH PART step",
 				},
+				cli.BoolFlag{
+					Name:   "restore-in-place",
+					Hidden: false,
+					Usage:  "Perform in-place restore by comparing backup parts with current database parts. Only downloads differential parts instead of full restore. Requires --data flag.",
+				},
 			),
 		},
 		{
@@ -544,7 +554,21 @@ func main() {
 			UsageText: "clickhouse-backup restore_remote [--schema] [--data] [-t, --tables=<db>.<table>] [-m, --restore-database-mapping=<originDB>:<targetDB>[,<...>]] [--tm, --restore-table-mapping=<originTable>:<targetTable>[,<...>]] [--partitions=<partitions_names>] [--rm, --drop] [-i, --ignore-dependencies] [--rbac] [--configs] [--named-collections] [--resumable] <backup_name>",
 			Action: func(c *cli.Context) error {
 				b := backup.NewBackuper(config.GetConfigFromCli(c))
-				return b.RestoreFromRemote(c.Args().First(), c.String("tables"), c.StringSlice("restore-database-mapping"), c.StringSlice("restore-table-mapping"), c.StringSlice("partitions"), c.StringSlice("skip-projections"), c.Bool("schema"), c.Bool("d"), c.Bool("rm"), c.Bool("i"), c.Bool("rbac"), c.Bool("rbac-only"), c.Bool("configs"), c.Bool("configs-only"), c.Bool("named-collections"), c.Bool("named-collections-only"), c.Bool("resume"), c.Bool("restore-schema-as-attach"), c.Bool("replicated-copy-to-detached"), c.Bool("hardlink-exists-files"), version, c.Int("command-id"))
+				// Override config with CLI flag if provided
+				if c.Bool("restore-in-place") {
+					b.SetRestoreInPlace(true)
+				}
+				// CI/CD Diagnostic: Log function signature validation for Go 1.25 + ClickHouse 23.8 compatibility
+				log.Debug().Fields(map[string]interface{}{
+					"operation":               "restore_remote_cli_validation",
+					"go_version":             runtime.Version(),
+					"hardlink_exists_files":  c.Bool("hardlink-exists-files"),
+					"drop_if_schema_changed": c.Bool("drop-if-schema-changed"),
+					"function_signature":     "RestoreFromRemote",
+					"parameter_count":        "expected_22_parameters",
+					"issue_diagnosis":        "hardcoded_false_should_be_cli_flag",
+				}).Msg("diagnosing CI Build/Test (1.25, 23.8) function signature mismatch")
+				return b.RestoreFromRemote(c.Args().First(), c.String("tables"), c.StringSlice("restore-database-mapping"), c.StringSlice("restore-table-mapping"), c.StringSlice("partitions"), c.StringSlice("skip-projections"), c.Bool("schema"), c.Bool("d"), c.Bool("rm"), c.Bool("i"), c.Bool("rbac"), c.Bool("rbac-only"), c.Bool("configs"), c.Bool("configs-only"), c.Bool("named-collections"), c.Bool("named-collections-only"), c.Bool("resume"), c.Bool("restore-schema-as-attach"), c.Bool("replicated-copy-to-detached"), c.Bool("hardlink-exists-files"), c.Bool("drop-if-schema-changed"), version, c.Int("command-id"))
 			},
 			Flags: append(cliapp.Flags,
 				cli.StringFlag{
@@ -642,6 +666,16 @@ func main() {
 					Name:   "hardlink-exists-files",
 					Hidden: false,
 					Usage:  "Create hardlinks for existing files instead of downloading",
+				},
+				cli.BoolFlag{
+					Name:   "restore-in-place",
+					Hidden: false,
+					Usage:  "Perform in-place restore by comparing backup parts with current database parts. Only downloads differential parts instead of full restore. Requires --data flag.",
+				},
+				cli.BoolFlag{
+					Name:   "drop-if-schema-changed",
+					Hidden: false,
+					Usage:  "Drop and recreate tables when schema changes are detected during in-place restore. Only available with --restore-in-place flag.",
 				},
 			),
 		},